feat: 새로운 ApplicationJpaRepository 및 ApplicationSchoolJpaRepository 추가
